lxd-to-incus: Split out validation code
